Extract evm client into separate package
