+++ analyzer/runtime/evm/client_test.go
@@ -1,4 +1,4 @@
-package runtime
+package evm
 
 import (
 	"context"
